Fetch snow and animations from modern minecraft.

+++ Launcher/Patcher/ResourceFetcher.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Drawing;
+using System.Drawing.Imaging;
 using System.IO;
 using System.Net;
 using System.Windows.Forms;
@@ -13,29 +14,43 @@ namespace Launcher {
 		const string classicJarUri = "http://s3.amazonaws.com/Minecraft.Download/versions/c0.30_01c/c0.30_01c.jar";
 		const string modernJarUri = "http://s3.amazonaws.com/Minecraft.Download/versions/1.6.2/1.6.2.jar";
 		const string terrainPatchUri = "http://static.classicube.net/terrain-patch.png";
-		static int resourcesCount = 2;
+		static int resourcesCount = 3;
 		
 		public void Run( MainForm form ) {
 			using( WebClient client = new GZipWebClient() ) {
 				WebRequest.DefaultWebProxy = null;
 				int i = 0;
 				DownloadData( classicJarUri, client, "classic.jar", form, ref i );
+				DownloadData( modernJarUri, client, "1.6.2.jar", form, ref i );
 				DownloadData( terrainPatchUri, client, "terrain-patch.png", form, ref i );
 			}
 			
 			reader = new ZipReader();
 			reader.ShouldProcessZipEntry = ShouldProcessZipEntry_Classic;
 			reader.ProcessZipEntry = ProcessZipEntry_Classic;
-			using( FileStream src = new FileStream( "classic.jar", FileMode.Open, FileAccess.Read, FileShare.Read ),
+			
+			using( FileStream srcClassic = File.OpenRead( "classic.jar" ),
+			      srcModern = File.OpenRead( "1.6.2.jar" ),
 			      dst = new FileStream( "default.zip", FileMode.Create, FileAccess.Write ) ) {
 				writer = new ZipWriter( dst );
-				reader.Extract( src );
+				reader.Extract( srcClassic );
+				
+				// Grab animations and snow
+				animBitmap = new Bitmap( 512, 32, PixelFormat.Format32bppArgb );
+				reader.ShouldProcessZipEntry = ShouldProcessZipEntry_Modern;
+				reader.ProcessZipEntry = ProcessZipEntry_Modern;
+				reader.Extract( srcModern );
+				writer.WriteNewImage( animBitmap, "animations.png" );
+				animBitmap.Dispose();
 				writer.WriteCentralDirectoryRecords();
 			}
-			File.Move( "terrain-patch.png", "terrain-patched.png" );
+			
+			if( !File.Exists( "terrain-patched.png" ) )
+				File.Move( "terrain-patch.png", "terrain-patched.png" );
 		}
 		ZipReader reader;
 		ZipWriter writer;
+		Bitmap animBitmap;
 		
 		bool ShouldProcessZipEntry_Classic( string filename ) {
 			return filename.StartsWith( "mob" ) || ( filename.IndexOf( '/' ) < 0 );
@@ -52,7 +67,29 @@ namespace Launcher {
 			using( Bitmap dstBitmap = new Bitmap( new MemoryStream( data ) ),
 			      maskBitmap = new Bitmap( "terrain-patch.png" ) ) {
 				PatchImage( dstBitmap, maskBitmap );
-				writer.WriteTerrainImage( dstBitmap, entry );
+				writer.WriteNewImage( dstBitmap, "terrain.png" );
+			}
+		}
+		
+		bool ShouldProcessZipEntry_Modern( string filename ) {
+			return filename.StartsWith( "assets/minecraft/textures" ) &&
+				( filename == "assets/minecraft/textures/environment/snow.png" ||
+				 filename == "assets/minecraft/textures/blocks/water_still.png" ||
+				 filename == "assets/minecraft/textures/blocks/lava_still.png" );
+		}
+		
+		void ProcessZipEntry_Modern( string filename, byte[] data, ZipEntry entry ) {
+			switch( filename ) {
+				case "assets/minecraft/textures/environment/snow.png":
+					entry.Filename = "snow.png";
+					writer.WriteZipEntry( entry, data );
+					break;
+				case "assets/minecraft/textures/blocks/water_still.png":
+					PatchAnimation( data, 0 );
+					break;
+				case "assets/minecraft/textures/blocks/lava_still.png":
+					PatchAnimation( data, 16 );
+					break;
 			}
 		}
 		
@@ -66,6 +103,20 @@ namespace Launcher {
 					for( int x = 0; x < size; x += tileSize ) {
 						if( row[x] != unchecked((int)0x80000000) ) {
 							FastBitmap.MovePortion( x, y, x, y, src, dst, tileSize );
+						}
+					}
+				}
+			}
+		}
+		
+		unsafe void PatchAnimation( byte[] data, int y ) {
+			// Sadly files in modern are 24 rgb, so we can't use fastbitmap here
+			using( Bitmap bmp = new Bitmap( new MemoryStream( data ) ) ) {			
+				for( int tile = 0; tile < bmp.Height; tile += 16 ) {
+					for( int yy = 0; yy < 16; yy++ ) {
+						for( int xx = 0; xx < 16; xx++ ) {
+							animBitmap.SetPixel( tile + xx, y + yy,
+							                    bmp.GetPixel( xx, tile + yy ) );
 						}
 					}
 				}
